brilliox-1/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
requests==2.31.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
openai==1.12.0
google-generativeai==0.3.2
duckduckgo-search==4.1.1
httpx==0.24.1
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
accept-paymob
starlette
